+++ Jesus/index.js
@@ -3,27 +3,7 @@ const AWS = require('aws-sdk');
 const docClient = new AWS.DynamoDB.DocumentClient();
 const tableName = process.env.TABLE_NAME;
 const { processMessage } = require('./processMessage');
-
-async function get(id) {
-    const data = await docClient.get({
-        TableName: tableName,
-        Key: {
-            'id': id
-        }
-    }).promise();
-
-    return data.Item;
-}
-
-async function upsert(id, item) {
-    await docClient.put({
-        TableName: tableName,
-        Item: {
-            ...item,
-            id: id,
-        }
-    }).promise();
-}
+const { get, upsert } = require('./db');
 
 async function addMembers(chatId, message) {
     const newChatMembers = message.new_chat_members;
